requirements.txt
# Core dependencies
streamlit>=1.28.0
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0